requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Cloud & AI
google-cloud-aiplatform==1.71.1
google-generativeai==0.3.2
vertexai==1.71.1

# ElevenLabs
elevenlabs==1.0.0

# Audio Processing
pydub==0.25.1
soundfile==0.12.1

# HTTP & WebSockets
httpx==0.26.0
websockets==12.0

# CORS & Security
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
